Flytt tidslinje extension function
